requirements.txt
requests
PyMuPDF
langchain-openai
langchain-core
PyYAML
python-dotenv
Pillow
pytesseract
fonttools